speed improvements to the rc4 cipher
git-svn-id: svn://svn.code.sf.net/p/axtls/code/trunk@107 9a5d90b5-6617-0410-8a86-bb477d3ed2e3

+++ ssl/crypto.h
@@ -64,7 +64,7 @@ void AES_convert_key(AES_CTX *ctx);
 
 typedef struct 
 {
-    int x, y, m[256];
+    uint8_t x, y, m[256];
 } RC4_CTX;
 
 void RC4_setup(RC4_CTX *s, const uint8_t *key, int length);
